moved service info into service package
