--- packages/orchestrator/internal/grpcserver/info_store.go
@@ -1,63 +1,0 @@
-package grpcserver
-
-import (
-	"context"
-	"time"
-
-	"github.com/google/uuid"
-
-	"github.com/e2b-dev/infra/packages/orchestrator/internal/sandbox"
-	"github.com/e2b-dev/infra/packages/orchestrator/internal/service"
-	orchestratorinfo "github.com/e2b-dev/infra/packages/shared/pkg/grpc/orchestrator-info"
-	"github.com/e2b-dev/infra/packages/shared/pkg/smap"
-)
-
-var (
-	serviceRolesMapper = map[service.ServiceType]orchestratorinfo.ServiceInfoRole{
-		service.Orchestrator:    orchestratorinfo.ServiceInfoRole_Orchestrator,
-		service.TemplateManager: orchestratorinfo.ServiceInfoRole_TemplateManager,
-	}
-)
-
-type Server struct {
-	orchestratorinfo.UnimplementedInfoServiceServer
-
-	info      *ServiceInfo
-	sandboxes *smap.Map[*sandbox.Sandbox]
-}
-
-func NewInfoService(_ context.Context, grpc *GRPCServer, info *ServiceInfo, sandboxes *smap.Map[*sandbox.Sandbox]) *Server {
-	service := &Server{
-		info:      info,
-		sandboxes: sandboxes,
-	}
-
-	orchestratorinfo.RegisterInfoServiceServer(grpc.GRPCServer(), service)
-
-	return service
-}
-
-func NewInfoContainer(clientId string, sourceVersion string, sourceCommit string) *ServiceInfo {
-	services := service.GetServices()
-	serviceRoles := make([]orchestratorinfo.ServiceInfoRole, 0)
-
-	for _, service := range services {
-		if role, ok := serviceRolesMapper[service]; ok {
-			serviceRoles = append(serviceRoles, role)
-		}
-	}
-
-	serviceInfo := &ServiceInfo{
-		ClientId:  clientId,
-		ServiceId: uuid.NewString(),
-		Startup:   time.Now(),
-		Roles:     serviceRoles,
-
-		SourceVersion: sourceVersion,
-		SourceCommit:  sourceCommit,
-	}
-
-	serviceInfo.SetStatus(orchestratorinfo.ServiceInfoStatus_OrchestratorHealthy)
-
-	return serviceInfo
-}
